+++ packages/ddp-client/client/client_convenience.js
@@ -1,6 +1,6 @@
 import { DDP } from '../common/namespace.js';
 import { Meteor } from 'meteor/meteor';
-import { loadAsyncStubHelpers } from "./queueStubsHelpers";
+import { loadAsyncStubHelpers } from "./queue_stub_helpers";
 
 // Meteor.refresh can be called on the client (if you're in common code) but it
 // only has an effect on the server.
